feat: landing find next talent section - move ProfileListItem jsx to ProfileCard component
